Moved the methods in UtilitySubsystem.java to other classes
