+++ src/main/java/dansplugins/spawnsystem/data/PersistentData.java
@@ -1,6 +1,9 @@
 package dansplugins.spawnsystem.data;
 
+import org.bukkit.ChatColor;
 import org.bukkit.Location;
+import org.bukkit.World;
+import org.bukkit.entity.Player;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -30,6 +33,33 @@ public class PersistentData {
 
     public ArrayList<UUID> getPlayersWithSpawns() {
         return playersWithSpawns;
+    }
+
+    public void setPlayersSpawn(Player player, World world, int x, int y, int z) {
+
+        Location spawnLocation = new Location(world, x, y, z);
+
+        // set spawn
+        if (!PersistentData.getInstance().getPlayerSpawns().containsKey(player.getName())) {
+            PersistentData.getInstance().getPlayerSpawns().put(player.getUniqueId(), spawnLocation);
+            PersistentData.getInstance().getPlayersWithSpawns().add(player.getUniqueId());
+        }
+        else {
+            player.sendMessage(ChatColor.RED + "You have already set your spawn! If you're starting a new character please see an admin for assistance.");
+            return;
+        }
+
+
+        // teleport player
+        player.teleport(spawnLocation);
+
+        player.sendMessage(ChatColor.GREEN + "Spawn set!");
+
+    }
+
+    public void resetSpawn(UUID player) {
+        PersistentData.getInstance().getPlayersWithSpawns().remove(player);
+        PersistentData.getInstance().getPlayerSpawns().remove(player);
     }
 
 }
