+++ src/main/java/dansplugins/spawnsystem/utils/BlockChecker.java
@@ -1,0 +1,39 @@
+package dansplugins.spawnsystem.utils;
+
+import org.bukkit.block.Block;
+
+public class BlockChecker {
+
+    private static BlockChecker instance;
+
+    private BlockChecker() {
+
+    }
+
+    public static BlockChecker getInstance() {
+        if (instance == null) {
+            instance = new BlockChecker();
+        }
+        return instance;
+    }
+
+    public boolean isSign(Block block) {
+        switch(block.getType()) {
+            case ACACIA_SIGN:
+            case ACACIA_WALL_SIGN:
+            case BIRCH_SIGN:
+            case BIRCH_WALL_SIGN:
+            case DARK_OAK_SIGN:
+            case DARK_OAK_WALL_SIGN:
+            case JUNGLE_SIGN:
+            case JUNGLE_WALL_SIGN:
+            case OAK_SIGN:
+            case OAK_WALL_SIGN:
+            case SPRUCE_SIGN:
+            case SPRUCE_WALL_SIGN:
+                return true;
+        }
+        return false;
+    }
+
+}
